feat(account creation): mv import progress into separate wizard step
