+++ js/src/constants/translations/en.json
@@ -269,6 +269,7 @@
     "iaso.snt_malaria.appName": "SNT Malaria Explorer",
     "iaso.snt_malaria.configureAccount.title": "Finish setting up your account",
     "iaso.snt_malaria.configureAccount.stepUserInfo": "Your information",
+    "iaso.snt_malaria.configureAccount.stepImportBoundaries": "Import",
     "iaso.snt_malaria.configureAccount.stepAccountSettings": "Account settings",
     "iaso.snt_malaria.configureAccount.stepBudgetSettings": "Budget settings",
     "iaso.snt_malaria.configureAccount.userInfoTitle": "Tell us about yourself",
@@ -282,6 +283,8 @@
     "iaso.snt_malaria.configureAccount.importDescription": "We are importing the boundaries from the GeoJSON file you uploaded. This usually takes less than a minute.",
     "iaso.snt_malaria.configureAccount.importErrored": "The import failed.",
     "iaso.snt_malaria.configureAccount.importKilled": "The import was cancelled.",
+    "iaso.snt_malaria.configureAccount.importSuccessTitle": "Import complete",
+    "iaso.snt_malaria.configureAccount.importSuccessDescription": "Your administrative boundaries have been imported successfully. Continue to the next step to configure your account.",
     "iaso.snt_malaria.configureAccount.configureTitle": "Account settings",
     "iaso.snt_malaria.configureAccount.configureDescription": "Pick the administrative levels we will use to draw your map and to plan interventions.",
     "iaso.snt_malaria.configureAccount.focusOrgUnitType": "Map zoom option",
