+++ js/src/constants/translations/fr.json
@@ -267,6 +267,7 @@
     "iaso.snt_malaria.appName": "SNT Malaria Explorer",
     "iaso.snt_malaria.configureAccount.title": "Finalisez la configuration de votre compte",
     "iaso.snt_malaria.configureAccount.stepUserInfo": "Vos informations",
+    "iaso.snt_malaria.configureAccount.stepImportBoundaries": "Import",
     "iaso.snt_malaria.configureAccount.stepAccountSettings": "Paramètres du compte",
     "iaso.snt_malaria.configureAccount.stepBudgetSettings": "Paramètres budgétaires",
     "iaso.snt_malaria.configureAccount.userInfoTitle": "Parlez-nous de vous",
@@ -280,6 +281,8 @@
     "iaso.snt_malaria.configureAccount.importDescription": "Nous importons les limites depuis le fichier GeoJSON que vous avez téléversé. Cela prend habituellement moins d'une minute.",
     "iaso.snt_malaria.configureAccount.importErrored": "L'import a échoué.",
     "iaso.snt_malaria.configureAccount.importKilled": "L'import a été annulé.",
+    "iaso.snt_malaria.configureAccount.importSuccessTitle": "Import terminé",
+    "iaso.snt_malaria.configureAccount.importSuccessDescription": "Vos limites administratives ont été importées avec succès. Passez à l'étape suivante pour configurer votre compte.",
     "iaso.snt_malaria.configureAccount.configureTitle": "Paramètres du compte",
     "iaso.snt_malaria.configureAccount.configureDescription": "Choisissez les niveaux administratifs utilisés pour afficher la carte et planifier les interventions.",
     "iaso.snt_malaria.configureAccount.focusOrgUnitType": "Option de zoom sur la carte",
